chore: programmatically validate that tool schemas do not use nullable (#189)
## Short description

In #181 we introduced instructions for AI coding agents to not use
`nullable` in tool arg schemas. This PR changes it to detect it
programmatically during linting checks. This will give AI coding tools
the feedback if they ever do it, without having to say so in the agent
instructions ("show, don't tell"). This also makes it deterministic that
we won't do it again.